requirements.txt
google-ads==22.0.0
google-cloud-storage==2.14.0
